Refactor: 책 가격 타입 Integer 로 변경

+++ src/main/java/com/usg/book/adapter/in/web/dto/BookRegisterRequest.java
@@ -10,11 +10,11 @@ public class BookRegisterRequest {
     private String bookName; // 책 이름
     private String bookComment; // 책 코멘트
     private String bookPostName; // 책 등록 시 게시글 이름
-    private Long bookPrice; // 책 가격
+    private Integer bookPrice; // 책 가격
     private String isbn; // ISBN
 
     @Builder
-    public BookRegisterRequest(Long memberId, String bookName, String bookComment, String bookPostName, Long bookPrice, String isbn) {
+    public BookRegisterRequest(Long memberId, String bookName, String bookComment, String bookPostName, Integer bookPrice, String isbn) {
         this.memberId = memberId;
         this.bookName = bookName;
         this.bookComment = bookComment;
